classes com herança

+++ src/main/java/br/edu/uncisal/shoppingcart/model/Product.java
@@ -1,7 +1,5 @@
 package br.edu.uncisal.shoppingcart.model;
 
-import java.math.BigDecimal;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -16,7 +14,7 @@ public class Product {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String name;
-	private BigDecimal price;
+	private Double price;
 	private Integer quantity;
 
 	public Long getId() {
@@ -35,11 +33,11 @@ public class Product {
 		this.name = name;
 	}
 
-	public BigDecimal getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 
@@ -53,7 +51,7 @@ public class Product {
 	
 	@Override
 	public String toString() {
-		return String.format("[%d %s %d %d]", this.id, this.name, this.price, this.quantity);
+		return String.format("[%d %s %f %d]", this.id, this.name, this.price, this.quantity);
 	}
 
 }
